backend/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
weaviate-client==3.24.1
sentence-transformers==2.2.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
numpy==1.26.2
pytest==7.4.3
httpx==0.25.2
